feat(#1535): CLEAN: split output.ts, four 100+ line functions
