+++ .pi/extensions/supervisor/agent/validation.ts
@@ -1,0 +1,148 @@
+// ─── Agent Output Validation ────────────────────────────────────
+// Schema validation for parsed agent output. Leaf module — must never
+// import from output.ts (ESM cycle risk); output.ts imports from here.
+
+import type { FindingSeverity } from "../config/types.ts";
+
+const VALID_SEVERITIES = new Set<FindingSeverity>(["critical", "warning", "suggestion"]);
+
+interface ValidationResult {
+	valid: boolean;
+	errors: string[];
+}
+
+function validateAgentOutput(data: Record<string, unknown>): ValidationResult {
+	const errors: string[] = [];
+
+	// action is required and must be a valid enum value
+	if (data.action === undefined || data.action === null) {
+		errors.push("Missing required field: 'action'");
+	} else if (typeof data.action !== "string") {
+		errors.push("'action' must be a string");
+	} else if (!["COMPLETE", "APPROVED", "REJECTED"].includes(data.action)) {
+		errors.push(`'action' must be one of: COMPLETE, APPROVED, REJECTED (got: ${data.action})`);
+	}
+
+	// agentName is required and must be a string
+	if (data.agentName === undefined || data.agentName === null) {
+		errors.push("Missing required field: 'agentName'");
+	} else if (typeof data.agentName !== "string") {
+		errors.push("'agentName' must be a string");
+	}
+
+	// refusal — if present, treat as rejection
+	if (data.refusal !== undefined && data.refusal !== null) {
+		if (typeof data.refusal === "string" && data.refusal.trim().length > 0) {
+			errors.push(`Agent refused: ${data.refusal}`);
+		}
+	}
+
+	// commentBody (optional, must be string if present)
+	if (
+		data.commentBody !== undefined &&
+		data.commentBody !== null &&
+		typeof data.commentBody !== "string"
+	) {
+		errors.push("'commentBody' must be a string if provided");
+	}
+
+	// prTitle (optional, must be string if present)
+	if (data.prTitle !== undefined && data.prTitle !== null && typeof data.prTitle !== "string") {
+		errors.push("'prTitle' must be a string if provided");
+	}
+
+	// prBody (optional, must be string if present)
+	if (data.prBody !== undefined && data.prBody !== null && typeof data.prBody !== "string") {
+		errors.push("'prBody' must be a string if provided");
+	}
+
+	// summary (optional, must be string if present)
+	if (data.summary !== undefined && data.summary !== null && typeof data.summary !== "string") {
+		errors.push("'summary' must be a string if provided");
+	}
+
+	// auditScore validation
+	if (data.auditScore !== undefined && data.auditScore !== null) {
+		if (typeof data.auditScore !== "object" || Array.isArray(data.auditScore)) {
+			errors.push("'auditScore' must be an object with 'passing' and 'total' fields");
+		} else {
+			const score = data.auditScore as Record<string, unknown>;
+			if (typeof score.passing !== "number" || typeof score.total !== "number") {
+				errors.push("'auditScore.passing' and 'auditScore.total' must be numbers");
+			} else {
+				if (score.passing < 0 || score.total < 0) {
+					errors.push("'auditScore.passing' and 'auditScore.total' must be non-negative");
+				}
+				if (score.passing > score.total) {
+					errors.push(
+						`'auditScore.passing' (${score.passing}) cannot exceed 'auditScore.total' (${score.total})`,
+					);
+				}
+			}
+		}
+	}
+
+	// findings validation
+	if (data.findings !== undefined && data.findings !== null) {
+		if (!Array.isArray(data.findings)) {
+			errors.push("'findings' must be an array if provided");
+		} else {
+			for (let i = 0; i < data.findings.length; i++) {
+				const f = data.findings[i];
+				if (typeof f !== "object" || f === null) {
+					errors.push(`findings[${i}] must be an object`);
+					continue;
+				}
+				const finding = f as Record<string, unknown>;
+
+				// severity
+				if (
+					typeof finding.severity !== "string" ||
+					!VALID_SEVERITIES.has(finding.severity as FindingSeverity)
+				) {
+					errors.push(
+						`findings[${i}].severity must be one of: ${Array.from(VALID_SEVERITIES).join(", ")}`,
+					);
+				}
+
+				// dimension
+				if (typeof finding.dimension !== "string") {
+					errors.push(`findings[${i}].dimension must be a string`);
+				}
+
+				// symptom, consequence, remedy are required strings
+				if (typeof finding.symptom !== "string" || finding.symptom.trim() === "") {
+					errors.push(`findings[${i}].symptom is required and must be a non-empty string`);
+				}
+				if (typeof finding.consequence !== "string" || finding.consequence.trim() === "") {
+					errors.push(`findings[${i}].consequence is required and must be a non-empty string`);
+				}
+				if (typeof finding.remedy !== "string" || finding.remedy.trim() === "") {
+					errors.push(`findings[${i}].remedy is required and must be a non-empty string`);
+				}
+
+				// location (optional)
+				if (
+					finding.location !== undefined &&
+					finding.location !== null &&
+					typeof finding.location !== "string"
+				) {
+					errors.push(`findings[${i}].location must be a string if provided`);
+				}
+			}
+		}
+	}
+
+	// targetStatus (optional, must be string if present)
+	if (
+		data.targetStatus !== undefined &&
+		data.targetStatus !== null &&
+		typeof data.targetStatus !== "string"
+	) {
+		errors.push("'targetStatus' must be a string if provided");
+	}
+
+	return { valid: errors.length === 0, errors };
+}
+
+export { validateAgentOutput, VALID_SEVERITIES };
